Wiring up the classes and Interfaces

+++ src/main/java/com/github/gurpreetsachdeva/creditcardsaggregatorservice/model/CreditCardUpstreamResponse.java
@@ -4,8 +4,8 @@ public class CreditCardUpstreamResponse {
 	
 			
 	private String cardName;
-	private float apr;
-	private float eligibility;
+	private double apr;
+	private double eligibility;
 	
 	public String getCardName() {
 		return cardName;
@@ -13,16 +13,16 @@ public class CreditCardUpstreamResponse {
 	public void setCardName(String cardName) {
 		this.cardName = cardName;
 	}
-	public float getApr() {
+	public double getApr() {
 		return apr;
 	}
-	public void setApr(float apr) {
+	public void setApr(double apr) {
 		this.apr = apr;
 	}
-	public float getEligibility() {
+	public double getEligibility() {
 		return eligibility;
 	}
-	public void setEligibility(float eligibility) {
+	public void setEligibility(double eligibility) {
 		this.eligibility = eligibility;
 	}
